+++ ontolib-io/src/main/java/com/github/phenomics/ontolib/io/obo/go/GoOboFactory.java
@@ -142,7 +142,7 @@ class GoOboFactory implements OboOntologyEntryFactory<GoTerm, GoTermRelation> {
     final String creationDateStr =
         (creationDateEntry == null) ? null : creationDateEntry.getValue();
 
-    final SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
+    final SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'");
     Date creationDate = null;
     if (creationDateStr != null) {
       try {